src/requirements.txt
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
argparse==1.2.1
mistune==0.7
pygments-github-lexers==0.01
wsgiref==0.1.2
